move trace fixtures into lib
